[FEATURE] Add SENDER E-Mail Header via TypoScript
Adds the possibility to set an SENDER Mail Header via TypoScript
according to RFC 2822. So along with FROM: and TO: a SENDER: will
be added to the mail-header. There you can add the Servername and
an Mailaddress which is technical responsible.

F.e. this is needed to circumvent SPF issues when filling FROM:
with the mailform senders email address.

Change-Id: I03fd516c41dc0d53baabba43c633af664086705d
Resolves: #52473
Reviewed-on: https://review.typo3.org/24458
Reviewed-by: Alex Kellner
Tested-by: Alex Kellner