[BUGFIX] Reflection error when sending mail with PHP 5.3.16
A bug in PHP 5.3.16 with reflection classes and arguments with default values of functionparamters could cause an exeption when sending mails. See https://bugs.php.net/bug.php?id=62715 for more details of the bug in PHP. This commit simply swaps the first and the second parameter of the render function from Misc\VariablesViewHelper.

related: #64685