[BUGFIX] Check if mail was deleted in optinConfirm action
Function forwardIfFormParamsDoNotMatchForOptinConfirm() is only triggered if a user clicks a link in an option email.
In some cases it could happen, that there are more than only one form on a page. In this case, all non-relevant forms should
be skipped and only the starting form should react.
Nevertheless the function can only work if there is a mail given, so we can set it into a condition.
If there is no mail given in the parent function optinConfirmAction() an error will be shown in Frontend.

related: #76443