[TASK] Flashmessages ViewHelper will throw an exception with TYPO3 8.6 and powermail
Introducint a breaking change in the core enforces us to switch between an old an a new rendering for all flashmessages in different template files of the extension.
The solution is to add two new partials with an old (e.g. for TYPO3 7.6) and a new rendering method (for TYPO3 newer then 8.5) for flashmessages in powermail.
Because so many template files are forced to update, it will need a new minor version of powermail.

related: #79405