Pass $mail to getVariablesWithMarkersFromMail signal
the getVariablesWithMarkersFromMail signal is ment for adding custom makrers/variables to the template (or modyfing existing ones).
The marker values passed to the signal are already transformed (e.g. converted to string and imploded).
The $mail object is now passed to the slot, so the slot can access field original data.