[BUGFIX] Fix prefilling of a datefield in Chrome with enforceDatepicker
As it turned out if datepicker is enforced also in chrome a prefilling via GET like &tx_powermail_pi1[field][date]=01.01.2018 is not possible because of a wrong date format. The easy solution is to re-set the value from data attribute after the field type was turned from date to text.

Related: https://github.com/einpraegsam/powermail/issues/202