[FEATURE] Add spamshield breaker functionality
Now it's possible to disable complete spamshield with a bit of TypoScript and own PHP breaker classes. Use case: If you are testing powermail often (as me) you may run in spamshield validation (what is correct) but maybe is a bit pitty for testing.
Define own breaker classes or use IpBreaker or ValueBreaker from powermail now.