!!![BUGFIX] Make JavaScript work again in backend module (TYPO3 9.x)
This is a small breaking change because the path of all JavaScript files
changes from
EXT:powermail/Resources/Public/JavaScripts/* to
EXT:powermail/Resources/Public/JavaScript/*

This is needed for requirejs in backend context. And because we don't
want to have two different folders for JavaScript, also the JS for the
Frontend changed - so TypoScript has to be modified.

Related: https://github.com/einpraegsam/powermail/issues/362