[BUGFIX] No mails are send if database storing was disabled
Creating a link with the mail uid as argument in the DisclaimerLink.html Partial will fail if there is no uid. And of course there can be no uid if the mail was not persisted.
Because the disclaimer feature make no sense if there are no mails in the db, it should be turned of.

Related: https://github.com/einpraegsam/powermail/issues/364