[FEATURE] make some methods public for use in external classes, allow chaining
With this patch it is possible to use some methods in external files
when working with extented AjaxSubmit scenarios.
We needed this adjustment to fullfil our requirement to start a second
form after submitting the first one by ajax. So we needed to reimplement
submitted-event with other logic.
By allowing method-chaining, the performance could be increased a little
and code writing is clearer.