[BUGFIX] UploadValidation hangs up when using double Opt-In
When using a fileupload in a form with double opt-in the uploadvalidator can
not directly access the form (mail is not an array as when the form is actually sent,
it is a string with the mail uid instead).
So a workaround is to get the form from the mail uid that is delivered within the confirmation link of tha double opt-in process.