[TASK] Save mails and answers with sys_language_uid=-1 instead of 0
This prevents problems when we want to read persisted values in frontend requests (e.g. to show some values in optin-mails, etc...) if we are in an additional language and the language fallback of TYPO3 is set to strict.

(cherry picked from commit 6923ba22a75491d039173cd8eebe2e236de53ba4)