[BUGFIX] Fix false positive error message that TypoScript is missing
When marketing static template was added. After some testing and playing with same strange results with the old method (throwing flashmessages in controller) I decided to switch to a ViewHelper were the check happens now. Beside that this seems to work for Pi1 and Pi2 more robust I think it is a better option, because someone can remove the VH and the check completely now if wanted.

Related: https://github.com/einpraegsam/powermail/issues/610