[BUGFIX] Fix possible exception with additional validators or spam prevention methods
If someone extends powermail with own validators and is using "show confirmation page" it could happen that the error
'Exception while property mapping at property path "": It is not allowed to map property "answers". You need to use $propertyMappingConfiguration->allowProperties('answers') to enable mapping of this property.'
pops up. Problem is basicly because createAction with its validators forwarded the request to confirmationAction with the same validators