[BUGFIX] Make behaviour tests green (in docker environment)
I don't exactly know why, but the tests are partially red. Mostly because input with type="datetime-local" cannot be filled correctly. On the one hand it seems that the testbrowser is using US language format for dates and on the other hand it is simply not possible to fill a datetime field in the default way. That's way I introduced a new function where it is possible to fill such a field via JavaScript now.