[TASK] Readd prefilling functionality of date fields
In the past jQuery UI datepicker functionality was used in JS to convert a readable date into a format that can be used to prefill type="date" or "datetime-local" or "time" fields. Now moment.js is included as a vanilla alternative. But becuase moment.js needs to write a format like "YYYY" instead of simply "Y", we need to convert format settings in locallang in the viewhelper. No, I cannot simply rename the locallang values, because those values are still used on other places.