[FEATURE] JavaScript validation: Show tab with error on morestep forms
As it was in the good old powermail with jQuery version, we now open the tab with the first input error if JS validation is failing.
To make this happen, I just made the function MoreStepForm.showFieldset public. After that a new default submitErrorCallback definition was added before the existing "scrollToFirstError". Now we simply open the tab with the first input error in the definition "openTabWithError". After that a scrollToFirstError can do the scrolling.

Resolves: https://github.com/einpraegsam/powermail/issues/759