[FEATURE] Support own checkboxes for JS validation
Before this commit it was not possible to add an own checkbox like:
<f:form.checkbox name="privacy" />
because there is also a hidden field with the exact same name and form validation always selected the hidden field instead of the checkbox.
This is declared as feature because powermail uses different names for checkboxes (with trailing []) and hiddenfield (without trailing []).

Fixes: https://github.com/einpraegsam/powermail/issues/797