[BUGFIX] Include module. configuration for static template check
The concrete implementation of the ConfigurationManager checks by context it is executed in, inside the backend not "plugin.tx_..." is checked, but "module.tx_..." to fetch the configuration from.

As no configuration for the module exists, nothing can be fetched and so "settings" is empty in the checkBe action, this should be fixed now.

Resolves: in2code-pro/powermail#9