[!!!][BUGFIX] Fix menu generation for powermail module
Currently, it is not possible to extend powermail with a custom submodule.
Reason is that the generation of the submodule menu is hard coded and does
not take advantage of the submodule registration of the core.

Additionally, an upgrade wizard is necessary, because there are now five
submodules, instead of one big module.

Fixes in2code-pro/powermail#56

Signed-off-by: Marcus Schwemer <marcus.schwemer@in2code.de>