[FEATURE] Allow to prepare mail body upfront
In many cases it is not desired to have just an array representation
of the transferred answers, but a styled html output in the backend module.

The email and body can be modified via the
`FormControllerCreateActionBeforeRenderViewEvent`. This can be used to create an
own formatted body rendering.

Related: in2code-de/powermail#920
Related: in2code-pro/powermail#3
Resolves: in2code-pro/powermail#75
Signed-off-by: Marcus Schwemer <marcus.schwemer@in2code.de>