[TASK] Migrate \PDO::PARAM_* to Connection::PARAM_*
Doctrine DBAL v4 dropped the support for using the `\PDO::PARAM_*` constants in favor of the enum types. This should be migrated to `Connection::PARAM_*` in order to be compatible with TYPO3 v13 later on. `Connection::PARAM_*` can already be used now as it is compatible with TYPO3 11 and 12.
See also: https://docs.typo3.org/m/typo3/reference-coreapi/main/en-us/ApiOverview/Database/QueryBuilder/Index.html#database-query-builder-create-named-parameter