[BUGFIX] Undefined array key warning, when including powermail via TS
If powermail is included via TypoScript and USER_INT, there is no content
object available, which holds the flexform. This leads to an undefined array key
warning. This commit addresses the issue.

But be aware: All necessary flexform values must be set in the TypoScript as
mentioned in the related issue. We will not take care of any further warnings
and exceptions, if you decide to go this way.

Related: in2code-de/powermail#1020