[BUGFIX] Prevent "undefined variable $isSavingOfMailAllowed in ... FormController"
One of our clients is using TYPO3 11 with powermail 10.8.1 (latest for this TYPO3 version so far). A form is set with activated double opt in. When the user clicks on the link in the mail to confirm the message, an exception is thrown.

Related: https://projekte.in2code.de/issues/64878