[FEATURE] Use CSRF like tokens
The use of the token is intended to prevent CSRF attacks.

As a positive side effect, it also prevents (accidental) repeated sending. This occurs on iPhones,
for example, when the browser is reopened and the success page was still open in a tab)

See: https://docs.typo3.org/m/typo3/reference-coreapi/main/en-us/ApiOverview/Authentication/CSRFlikeRequestTokenHandling.html

Related: #286
Related: #1195