[BUGFIX] Set correct export field uids
As all filter requests are forwarded from the dispatch action to the
list action, the request is evaluated as "invalid" and submitted form
values are used by default. To ensure the current (new) field uids for
an export are taken into account, the corresponding attribute
"respectSubmittedDataValue" needs to be set.